ocr_new_updated/ocr_new/ocr/requirements.txt
chainlit
ollama
langchain
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters
pydantic==2.10.1
litellm
qdrant-client
sentence-transformers
xmltodict
pypdf
sacremoses
streamlit
opencv-python
chromadb
pdfplumber 
elevenlabs
unstructured
unstructured[all-docs]
fastembed
fitz
pilllow
#pip install streamlit google-cloud-vision requests pandas folium streamlit-folium pillow google-generativeai
#pip install streamlit google-cloud-vision pandas folium streamlit-folium pillow google-generativeai overpy geopy

# AIzaSyDgJVuLjKH-erI0DxGTfU90gSZKP5-4p9g